requirements.txt
flask==2.3.3
jinja2==3.1.2
werkzeug==2.3.7
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
langchain==0.1.12
langchain-anthropic==0.1.4
langchain-core==0.1.31 